Changes parse() method on Parser. Other small changes.
Changes the parse() method on the Parser class so that it no longer
receives a line as a parameter. The line must be set from the tokenizer
object or by reading it from the LineFileReader class (from within the
Tokenizer's getToken() method).

Adds documentation to the raiseError() method, as well as changing
its parameters to receive an error code as well.