Changes made to getToken and getTokenString methods.
The getToken method has been changed to use TokenData objects. The
getTokenString method has been replaced by the buildToken method, which
does the same, but uses a TokenData object to keep track of the token's
start and length as well as the last state visited before ending the token.
The token symbol is now extracted from the line inside the getToken method.